requirements.txt
firebase_admin==6.6.0
joblib==1.4.2
numpy==2.3.3
pandas==2.3.2
psycopg2==2.9.10
python-dotenv==1.1.1
reportlab==4.3.1
Requests==2.32.5
streamlit==1.41.1
streamlit_option_menu==0.4.0
supabase==2.18.1
